+++ src/components/common/ProgressBarWithDataCount/ProgressBarWithDataCount.js
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import ProgressBar from 'common/ProgressBar';
 import { goalNum } from '../../../constants/dataProgress';
 import { queryExperienceCountIfUnfetched as queryExperienceCount } from '../../../actions/experiences';
-import { experienceCountSelector } from '../../../containers/PermissionBlock/selectors';
+import { experienceCountSelector } from '../../../selectors/countSelector';
 
 const ProgressBarWithDataCount = props => {
   const experienceCount = useSelector(experienceCountSelector);
