+++ src/containers/PermissionBlock/selectors.js
@@ -1,29 +1,16 @@
 import R from 'ramda';
 
 import { isFetched } from '../../constants/status';
-import {
-  menuEntriesSelector,
-  menuStatusSelector,
-} from '../../selectors/laborRightsSelector';
-
-export const experienceCountSelector = state => state.experiences.get('count');
+import { menuStatusSelector } from '../../selectors/laborRightsSelector';
 
 export const hasFetchedexperienceCountSelector = R.compose(
   isFetched,
   state => state.experiences.get('countStatus'),
 );
 
-export const timeAndSalaryCountSelector = state =>
-  state.timeAndSalary.get('count');
-
 export const hasFetchedTimeAndSalaryCountSelector = R.compose(
   isFetched,
   state => state.timeAndSalary.get('countStatus'),
-);
-
-export const laborRightsCountSelector = R.compose(
-  R.length,
-  menuEntriesSelector,
 );
 
 export const hasFetchedLaborRightsCountSelector = R.compose(
